requirements.txt
aiolimiter==1.1.0
annotated-types==0.6.0
anyio==4.2.0
APScheduler==3.10.4
cachetools==5.3.2
certifi==2023.11.17
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
pydantic==2.7.0
pydantic_core==2.18.1
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.29
sqlmodel==0.0.16
tornado==6.3.3
typing_extensions==4.11.0
tzlocal==5.2
